Added task name in communication
Added in sockets_overlay a procedure to found the task_id in the xml string in function of the event type.

Works but this current way of getting scheduling information via xml is fucked up.  Should find a way to get access directly to pure String values in the next steps.